move InternalErr to the syserr package
